backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
openai-agents>=0.0.15
openai>=1.76.0
qdrant-client>=1.11.0
asyncpg>=0.29.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
pydantic>=2.10
pydantic-settings>=2.5.0
python-dotenv>=1.0.1
httpx>=0.27.0
python-multipart>=0.0.12
